requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pillow>=10.0.0
tqdm>=4.66.0

# Optional geospatial (comment if not needed)
rasterio>=1.3.9
geopandas>=0.14.0
shapely>=2.0.0